aider: Add start script to package.json for React front-end server.
# Aider chat conversation:

ASSISTANT: To add a start script to the `package.json` for the React front-end server, I will add the `"start": "react-scripts start"` line to the `scripts` object. Here is the diff for the change:

```diff
--- client/package.json
+++ client/package.json
@@ ... @@
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "react-scripts start"
   },
@@ ... @@
```

This change will allow you to start the React development server by running `npm start` from the command line within the `client` directory.