aider: Create the `index.css` file with basic styles for the React project.
# Aider chat conversation:

USER: You also need to add an index.css file since you are importing it.

ASSISTANT: To create the `index.css` file that is being imported in the `client/src/index.js` file, I will provide a diff for a new file. This CSS file will be placed in the `src` directory of the React project.

Here is the diff for creating the `index.css` file:

```diff
--- /dev/null
+++ client/src/index.css
@@ ... @@
+/* Add your styles here */
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
+}
```

This `index.css` file includes some basic styles to normalize the body margin and set a default font family, which are common practices in web development. You can expand upon this with additional styles as needed for your application.